Move type function functions to Type from Argument
