+++ src/struct/commands/arguments/Argument.js
@@ -551,75 +551,6 @@ class Argument {
         if (phrase) return phrase;
         return null;
     }
-
-    /**
-     * Creates a type from multiple types (union type).
-     * The first type that resolves to a non-void value is used.
-     * @param {...ArgumentType|ArgumentTypeFunction} types - Types to use.
-     * @returns {ArgumentTypeFunction}
-     */
-    /* eslint-disable no-invalid-this */
-    static union(...types) {
-        return async function typeFn(phrase, message, args) {
-            for (let entry of types) {
-                if (typeof entry === 'function') entry = entry.bind(this);
-                // eslint-disable-next-line no-await-in-loop
-                const res = await Argument.cast(entry, this.handler.typeHandler, phrase, message, args);
-                if (res != null) return res;
-            }
-
-            return null;
-        };
-    }
-
-    /**
-     * Creates a type from multiple types (tuple type).
-     * Only inputs where each type resolves with a non-void value are valid.
-     * @param {...ArgumentType|ArgumentTypeFunction} types - Types to use.
-     * @returns {ArgumentTypeFunction}
-     */
-    static tuple(...types) {
-        return async function typeFn(phrase, message, args) {
-            const results = [];
-            for (const entry of types) {
-                // eslint-disable-next-line no-await-in-loop
-                const res = await Argument.cast(entry, this.handler.typeHandler, phrase, message, args);
-                if (res == null) return null;
-                results.push(res);
-            }
-
-            return results;
-        };
-    }
-
-    /**
-     * Creates a type with extra validation.
-     * If the predicate is not true, the value is considered invalid.
-     * @param {ArgumentType|ArgumentTypeFunction} type - The type to use.
-     * @param {ArgumentPredicate} predicate - The predicate function.
-     * @returns {ArgumentTypeFunction}
-     */
-    static validate(type, predicate) {
-        return async function typeFn(phrase, message, args) {
-            const res = await Argument.cast(type, this.handler.typeHandler, phrase, message, args);
-            if (res == null) return null;
-            if (!predicate(res, phrase, message, args)) return null;
-            return res;
-        };
-    }
-    /* eslint-enable no-invalid-this */
-
-    /**
-     * Creates a type where the parsed value must be within a range.
-     * @param {ArgumentType|ArgumentTypeFunction} type - The type to use
-     * @param {number} min - Minimum value.
-     * @param {number} max - Maximum value.
-     * @param {boolean} [inclusive=false] - Whether or not to be inclusive on the upper bound.
-     * @returns {ArgumentTypeFunction}
-     */
-    static range(type, min, max, inclusive = false) {
-        return Argument.validate(type, x => x >= min && (inclusive ? x <= max : x < max));
-    }
 }
 
 module.exports = Argument;
